WIP: Protect Access to Rosetta
In this commit, we add functionality to prevent a user from
accessing Rosetta if they aren't a team member or trusted
translator.

On the client, we get website user data from
`/services/check-login`.

On the server, we set up a route for mock website user data
if the user is developing locally.