perf: Measure API Perf., Add WIP Rewrite
We add `console.time()`'s to measure performance of
`getTranslationFormData` and other API functions. We discover that the
bottleneck in the API is the `getTranslationFormData` function. We then
discover that the bottleneck in the `getTranslationFormData` function is
the `getCommonTranslatedStringKeysAndStrings` function. We add a WIP
rewrite of the `getCommonTranslatedStringKeysAndStrings` function with
the aim to improve the speed of the API.