fix: ensure shared strings are committed
Basically, what happened here is that I needed to have a repo associated
with each shared string object in the translation form data. Once we had
a repo associated with each shared string object in the translation form
data, we were able to iterate over a comprehensive list of repos whose
strings were in the translation to create file contents for long-term
storage.

For https://github.com/phetsims/rosetta/issues/360.