feat: make inputs textarea, align copy value button
For https://github.com/phetsims/rosetta/issues/349. The original mission
for this commit was to make the inputs for long English strings be text
areas, but I think making them all text areas is a more elegant
solution.

I also aligned the copy value button and the text area using Flexbox
since it looked even weirder with the old styling and the textareas.