fix: make user ID string, await validation functions
For https://github.com/phetsims/rosetta/issues/368. JB and I were
testing the trigger build functionality prior to deploying Rosetta 2.0.
We noticed that the trigger build route wasn't working. There were
multiple issues with it that I fixed in this commit.

I also fixed the styling of the trigger build form, and made it so that
team members can see a link to the admin page.