chore: bump version number to 2.0.7
I was following a versioning scheme where each bug fix incremented the
patch number, but this was too fussy. Now, I increment the patch number
when I have a handful of bug fixes and small features ready to deploy.
Now the Git tags are a bit more sane. Hopefully if someone is confused
as to why the version number went down in the future, they will see this
commit message.